Add helpers methods
